city ​​DTO creation
