+++ patricare-back/src/modules/city/dto/city.dto.ts
@@ -1,15 +1,8 @@
-import { IsString } from 'class-validator';
-
-export class CreateCityDto {
-  @IsString() nom: string;
-  @IsString() codePostal: string;
-  @IsString() departement: string;
-  @IsString() region: string;
-}
-
-export class UpdateCityDto {
-  @IsString() nom?: string;
-  @IsString() codePostal?: string;
-  @IsString() departement?: string;
-  @IsString() region?: string;
+// dto/lodging.dto.ts
+export class cityDto {
+  id: number;
+  nom: string;
+  codePostal: string;
+  departement: string;
+  region: string;
 }
