+++ patricare-back/src/modules/city/dto/create-city.dto.ts
@@ -1,0 +1,8 @@
+import { IsString } from 'class-validator';
+
+export class CreateCityDto {
+  @IsString() nom: string;
+  @IsString() codePostal: string;
+  @IsString() departement: string;
+  @IsString() region: string;
+}
